requirements.txt
# Core dependencies
python-dateutil>=2.8.2
pyyaml>=6.0

# Web interface
flask>=3.0.0
requests>=2.31.0

# TUI interface
rich>=13.7.0
textual>=0.47.0

# Utilities
faker>=20.0.0
